+++ frontend/src/Assets/VEHICLE_ISSUES_DATA.ts
@@ -1,5 +1,6 @@
 import {
   ProjectTypes,
+  VehicleDataType,
   VehicleIssuesType,
 } from "../components/VehicleIssues/Store/VehicleIssuesTypes.tsx";
 
@@ -55,9 +56,118 @@ export const EXAMPLE_VEHICLE_ISSUES_DATA: VehicleIssuesType[] = [
       bodyNo: "GTW 123456",
       vehicleStage: "LP1",
     },
-    picture: "",
+    picture: "/EXAMPLE_PIC.jpeg",
     problemName: "Some issue on vehicle",
     problemDescription: "Description in detail for vehicle issue.",
     DTCCodes: ["P1000", "P2500"],
+  },
+  {
+    id: 2,
+    item: 2,
+    vehicleId: {
+      id: 2,
+      carIdentification: "C02",
+      bodyNo: "GTW 123458",
+      vehicleStage: "LP1",
+    },
+    picture: "/EXAMPLE_PIC.jpeg",
+    problemName: "Some issue on vehicle",
+    problemDescription: "Description in detail for vehicle issue.",
+    DTCCodes: ["P1000", "P2500"],
+  },
+  {
+    id: 3,
+    item: 3,
+    vehicleId: {
+      id: 3,
+      carIdentification: "C03",
+      bodyNo: "GTW 123459",
+      vehicleStage: "LP1",
+    },
+    picture: "/EXAMPLE_PIC.jpeg",
+    problemName: "Some issue on vehicle",
+    problemDescription: "Description in detail for vehicle issue.",
+    DTCCodes: ["P1000", "P2500"],
+  },
+  {
+    id: 4,
+    item: 4,
+    vehicleId: {
+      id: 1,
+      carIdentification: "C01",
+      bodyNo: "GTW 123456",
+      vehicleStage: "LP1",
+    },
+    picture: "/EXAMPLE_PIC.jpeg",
+    problemName: "Some issue on vehicle",
+    problemDescription: "Description in detail for vehicle issue.",
+    DTCCodes: ["P1000", "P2500"],
+  },
+  {
+    id: 5,
+    item: 5,
+    vehicleId: {
+      id: 2,
+      carIdentification: "C02",
+      bodyNo: "GTW 123458",
+      vehicleStage: "LP1",
+    },
+    picture: "/EXAMPLE_PIC.jpeg",
+    problemName: "Some issue on vehicle",
+    problemDescription: "Description in detail for vehicle issue.",
+    DTCCodes: ["P1000", "P2500"],
+  },
+  {
+    id: 6,
+    item: 6,
+    vehicleId: {
+      id: 3,
+      carIdentification: "C03",
+      bodyNo: "GTW 123459",
+      vehicleStage: "LP1",
+    },
+    picture: "/EXAMPLE_PIC.jpeg",
+    problemName: "Some issue on vehicle",
+    problemDescription: "Description in detail for vehicle issue.",
+    DTCCodes: ["P1000", "P2500"],
+  },
+];
+
+export const EXAMPLE_VEHICLE_DATA: VehicleDataType[] = [
+  {
+    id: 1,
+    carIdentification: "C01",
+    bodyNo: "GTW 123456",
+    vehicleStage: "LP1",
+  },
+  {
+    id: 2,
+    carIdentification: "C02",
+    bodyNo: "GTW 123458",
+    vehicleStage: "LP1",
+  },
+  {
+    id: 3,
+    carIdentification: "C03",
+    bodyNo: "GTW 123459",
+    vehicleStage: "LP1",
+  },
+  {
+    id: 4,
+    carIdentification: "C01",
+    bodyNo: "GTW 122222",
+    vehicleStage: "LP2",
+  },
+  {
+    id: 5,
+    carIdentification: "C02",
+    bodyNo: "GTW 333333",
+    vehicleStage: "LP2",
+  },
+  {
+    id: 6,
+    carIdentification: "C02",
+    bodyNo: "GTW 444444",
+    vehicleStage: "LP2",
   },
 ];